+++ pkg/hyperstart/mock/misc.go
