+++ pkg/hyperstart/mock/rand.go
